(Refactor) : reward 값 long 타입으로 변환

+++ src/main/java/lotto/model/Revenue.java
@@ -5,11 +5,11 @@ import java.util.List;
 public class Revenue {
     private int reward = 0;
 
-    private int oneRevenue(WinDetail winDetail){
+    private long oneRevenue(WinDetail winDetail){
         return winDetail.reward();
     }
 
-    public int revenue(List<WinDetail> result){
+    public long revenue(List<WinDetail> result){
         for(WinDetail win : result) {
             reward += oneRevenue(win);
         }
